Extracted cli scripts
