+++ conf/script/build-system/meson/setup/cli_fetch_compiler_version.py
@@ -1,0 +1,5 @@
+import colorama
+
+
+def cli_init():
+    colorama.init()
